+++ src/Sheet.php
@@ -352,6 +352,10 @@ class Sheet
 
             $row = $row->toArray($nullValue, $calculateFormulas, $formatData, $endColumn);
 
+            if (method_exists($import, 'isEmptyWhen') && $import->isEmptyWhen($row)) {
+                continue;
+            }
+
             if ($import instanceof WithMapping) {
                 $row = $import->map($row);
             }